Perzisztencia

Closes #107 and #108

See merge request szofttech-ab-2026/group-11/mav-csoport!75

+++ TransportTycoon/TransportTycoon.Model/TransportTycoonEventArgs.cs
@@ -3,13 +3,13 @@
     public sealed class TransportTycoonEventArgs : EventArgs
     {
         #region Fields
-        private readonly int _gameTime;
+        private readonly ulong _gameTime;
         private readonly int _maintance;
         private readonly int _numberOfVehicles;
         #endregion
 
         #region Properties
-        public int GameTime => _gameTime;
+        public ulong GameTime => _gameTime;
 
         public int Maintance => _maintance;
 
@@ -17,7 +17,7 @@
         #endregion
 
         #region Constructors
-        public TransportTycoonEventArgs(int gameTime, int numberOfVehicles, int maintance)
+        public TransportTycoonEventArgs(ulong gameTime, int numberOfVehicles, int maintance)
         {
             this._gameTime = gameTime;
             this._numberOfVehicles = numberOfVehicles;
